create interface classes
